bb_agent_requirements.txt
ipykernel
python-dotenv
openbb==4.1.7
# openbb-charting
selenium
openai
tiktoken
pydantic # ==1.10
langchain >= 0.2
langchain-openai
langchain-community
wikipedia
edgartools
sec-api
sec-parser
streamlit
